# api/api.yaml
openapi: 3.0.0
info:
  title: ZidiBackend API
  version: 1.0.0
  description: API for managing organizations and users in ZidiBackend
paths:
  /organizations:
    post:
      summary: Create an organization
      operationId: CreateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: Get all organizations
      operationId: GetOrganizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /organizations/{organization_id}:
    get:
      summary: Get an organization by ID
      operationId: GetOrganizationById
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update an organization
      operationId: UpdateOrganization
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete an organization
      operationId: DeleteOrganization
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Organization deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organization_id}/users:
    get:
      summary: Get all users in an organization
      operationId: GetUsersByOrganization
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of users in the organization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/search:
    get:
      summary: Get an organization by name
      operationId: GetOrganizationByName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"
  /users:
    post:
      summary: Create a user
      operationId: CreateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequestBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"

    get:
      summary: Get all users
      operationId: GetUsers
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/email:
    get:
      summary: Get a user by email
      operationId: GetUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}:
    get:
      summary: Get a user by ID
      operationId: GetUserById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update a user
      operationId: UpdateUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a user
      operationId: DeleteUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/password:
    put:
      summary: Update user password
      operationId: UpdateUserPassword
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                  format: password
                  description: The current password of the user.
                new_password:
                  type: string
                  format: password
                  description: The new password to set.
      responses:
        "200":
          description: Password updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated successfully.
        "400":
          description: Bad request (e.g., missing fields, weak password).
        "401":
          description: Unauthorized - Incorrect old password.
        "404":
          description: User not found.
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    Organization:
      type: object
      required:
        - id
        - email
        - contact_person_name
        - contact_person_phone
        - address
        - industry
        - company_size
        - company_name
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        contact_person_name:
          type: string
        contact_person_phone:
          type: string
        address:
          type: string
        industry:
          type: string
        company_size:
          type: integer
        company_name:
          type: string
...
    User:
      type: object
      required:
        - id
        - firstname
        - lastname
        - email
        - address
        - password
        - role
        - organization_id
      properties:
        id:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
        organization_id:
          type: string
          format: uuid
...
    CreateUserRequestBody:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phoneNumber
        - address
        - role
        - organization_id
        - password
        - confirmPassword
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
          format: password
        address:
          type: string
        confirmPassword:
          type: string
          format: password
        role:
          type: string
        organization_id:
          type: string
          format: uuid
...
    Error:
      type: object
      required:
        - message
        - error_code
      properties:
        error_code:
          type: string
        message:
          type: string
  responses:
    BadRequestError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"